mudança no campo data

+++ src/main/java/br/com/cliente/spring/service/ClienteService.java
@@ -1,6 +1,7 @@
 package br.com.cliente.spring.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,7 +22,14 @@ public class ClienteService {
 		return clienteRepository.findAll();
 	}
 	
-	public void deletaCliente(Long id) {
-		clienteRepository.deleteById(id);
+	public Cliente deletaCliente(Long id) {
+		Optional<Cliente> clienteOptional = clienteRepository.findById(id);
+		if (clienteOptional.isPresent()) {
+	        Cliente cliente = clienteOptional.get();
+	        clienteRepository.deleteById(id);
+	        return cliente;
+	    } else {
+	        throw new RuntimeException("Cliente não encontrado com o ID: " + id);
+	    }
 	}
 }
